+++ tests/integration/cron/audit_cleanup_validation_test.go
@@ -6,7 +6,6 @@ package cron_test
 import (
 	"bytes"
 	"encoding/json"
-	"fmt"
 	"net/http"
 	"testing"